--- UITests/Screens/SMT.swift
@@ -1,41 +1,0 @@
-//
-//  SMT.swift
-//  Ahavat Yeshua
-//
-//  Created by Ilya Sheynblat on 8/3/25.
-//
-
-//class SMT: BaseScreen {
-//    // MARK: - Given
-//    @discardableResult
-//    func givenIDoSMT() -> Self {
-//        return self
-//    }
-//    
-//    @discardableResult
-//    func givenIDoSMTElse() -> Self {
-//        return self
-//    }
-//    
-//    // MARK: - Then
-//    @discardableResult
-//    func thenIAssertSMT() -> Self {
-//        return self
-//    }
-//    
-//    @discardableResult
-//    func thenIAssertSMTElse() -> Self {
-//        return self
-//    }
-//    
-//    // MARK: - When
-//    @discardableResult
-//    func whenIDoSMT() -> Self {
-//        return self
-//    }
-//    
-//    @discardableResult
-//    func whenIDoSMTElse() -> Self {
-//        return self
-//    }
-//}
