+++ UITests/UITests/ProfileUITests.swift
@@ -82,18 +82,3 @@ extension ProfileUITests {
             .thenBannerTextShowDefault()
     }
 }
-//    TO DO:HOMEWORK
-//     create manual TCs for email field:
-//    
-//    positiveTC: test@gmail.com
-//    negativeTC: empty field + enter
-//     white space field
-//     white space between letters in front,back , in domain(3)
-//    missing @, .,(2)
-//    special characters #, $,% in front,back , in domain(3)
-//    numbers in front,back , in domain (3)
-//     upper cases only,in front,back, in domain (4)
-//     Camel  type cases in front,back, in domain (3)
-//     max 256 characters in email field
-    
-    // total 22 tests
